fix(wizard): BookingSidebar — date intere + typography dati chiave
Leggibilità:
- formatDate: month 'short' → 'long' ("11 mag" → "11 maggio")
- .booking-sidebar override: label da --text-xs a --text-sm (12→13px),
  value da --text-sm a --text-base (13→15px). Override contestuale,
  non impatta altri usi di .label-row-between.

Niente nuove classi: sfrutto specificity CSS dentro .booking-sidebar.

+++ components/wizard/BookingSidebar.tsx
@@ -27,7 +27,7 @@ function calcNights(ci: string, co: string) {
 function formatDate(ymd: string, locale: string) {
   return new Date(ymd + 'T00:00:00').toLocaleDateString(
     locale === 'it' ? 'it-IT' : locale === 'de' ? 'de-DE' : locale === 'pl' ? 'pl-PL' : 'en-GB',
-    { day: 'numeric', month: 'short' }
+    { day: 'numeric', month: 'long' }
   );
 }
